pyproject.toml
[project]
name = "expenses-ai-agent"
version = "0.1.0"
description = "An Expenses AI-assisted classification agent"
readme = "README.md"
authors = [
    { name = "Juan José Expósito González", email = "expositogonzalez.juanjose@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.68.0",
    "groq>=0.31.1",
    "openai>=1.109.1",
    "pydantic>=2.11.9",
    "python-decouple>=3.8",
    "python-telegram-bot>=21.11.0",
    "rich>=14.1.0",
    "sqlmodel>=0.0.25",
    "typer>=0.20.0",
    # Week 5: Web Interface
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "streamlit>=1.40.0",
    "plotly>=5.24.0",
    "pandas>=2.2.0",
]

[project.scripts]
expenses-ai-agent = "expenses_ai_agent:main"
expenses-telegram-bot = "expenses_ai_agent.telegram.bot:main"
expenses-api = "uvicorn:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "logfire>=4.10.0",
    "poethepoet>=0.37.0",
    "pytest-asyncio>=0.25.0",
    "pytest-coverage>=0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.6.0",
    "ty>=0.0.1a21",
]

[tool.poe.tasks]
test = "pytest -vv --cov=src"
test-integration = "pytest -vv -m integration"
test-validate = "pytest --cov=src --cov-fail-under=95"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests that require real API calls (deselect with '-m \"not integration\"')",
]
